Fix: compile the library and tests clean under the strict warning gate.
- Make all headers and unit tests warning-clean under -Wall -Wextra -Wconversion -Wsign-conversion -Wold-style-cast -Wdouble-promotion -Wshadow with -Werror, unblocking the Linux/macOS/Windows CI build that had never passed.
- Numerically neutral: integer index/size type hygiene only -- Eigen::Index for Eigen extents, static_cast<std::size_t> for standard-container indexing, explicit narrowing casts where a small dimension lands in an int field.
- Drop genuinely dead locals and unused parameters flagged by -Wunused-*, rename shadowing variables, and replace a tautological field-touch and a signed/unsigned comparison in tests.
- All 759 Debug unit tests pass unchanged.

+++ lib/argmin/include/argmin/derivative/finite_difference.h
@@ -32,7 +32,7 @@ void fd_gradient(const Problem& p,
                  const Eigen::Vector<Scalar, N>& x,
                  Eigen::Vector<Scalar, N>& g)
 {
-    const int n = x.size();
+    const int n = static_cast<int>(x.size());
     g.resize(n);
     const Scalar cbrt_eps = std::cbrt(std::numeric_limits<Scalar>::epsilon());
     Eigen::Vector<Scalar, N> x_perturbed = x;
@@ -70,7 +70,7 @@ void fd_jacobian(const VectorFunction& f,
                  Eigen::MatrixX<Scalar>& J,
                  int num_outputs)
 {
-    const int n = x.size();
+    const int n = static_cast<int>(x.size());
     J.resize(num_outputs, n);
     const Scalar cbrt_eps = std::cbrt(std::numeric_limits<Scalar>::epsilon());
     Eigen::Vector<Scalar, N> x_perturbed = x;
